Types in the right places
